+++ native/c/commands/ds.hpp
@@ -22,6 +22,7 @@ int handle_data_set_create_loadlib(InvocationContext &result);
 int handle_data_set_view(InvocationContext &result);
 int handle_data_set_list(InvocationContext &result);
 int handle_data_set_list_members(InvocationContext &result);
+int handle_data_set_resolve_alias(InvocationContext &result);
 int handle_data_set_write(InvocationContext &result);
 int handle_data_set_delete(InvocationContext &result);
 int handle_data_set_restore(InvocationContext &result);